+++ tools/Profiler/NanoProfiler/CLRProfiler/SummaryView.xaml
@@ -1,4 +1,4 @@
-﻿<Window x:Class="nanoFramework.Tools.NanoProfiler.CLRProfiler.SummaryForm"
+﻿<Window x:Class="nanoFramework.Tools.NanoProfiler.CLRProfiler.SummaryView"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
