+++ tools/Profiler/NanoProfiler/NanoProfiler.csproj
@@ -70,4 +70,8 @@
     </Compile>
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="NewFolder\" />
+  </ItemGroup>
+
 </Project>
